Moved game functions to another file
