+++ old/game_interaction.js
